a2a_server/requirements.txt
flask==2.3.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
google-generativeai==0.8.3